Moved :truck: crates/api -> src/
Issue: #41
