+++ Cargo.toml
@@ -3,21 +3,28 @@ name = "api-ittoku-tech"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-bench = false
-path = "crates/api/main.rs"
-name = "api-ittoku-tech"
-
-[workspace]
-members = ["crates/db"]
-
 [dependencies]
-db = { path = "crates/db" }
-
 actix-web = "4.3.1"
+chrono = "0.4.24"
 env_logger = "0.10.0"
 log = "0.4.17"
+serde = "1.0.162"
 serde_json = "1.0.96"
+
+[dependencies.sea-orm]
+version = "0.11.3"
+features = [
+    "debug-print",
+    "runtime-async-std-native-tls",
+    "sqlx-postgres",
+]
+
+[dependencies.sea-orm-migration]
+version = "0.11.0"
+features = [
+    "runtime-async-std-native-tls",
+    "sqlx-postgres",
+]
 
 # actix-web = "4.3.1"
 # chrono = { version = "0.4.24", features = ["serde"] }
